clean code reformat delete class QueryHandler merge them ti SE
