+++ src/main/java/search_engine/SearchEngine.java
@@ -1,5 +1,6 @@
 package search_engine;
 
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import search_engine.filters.Filter;
 import search_engine.queryDecoders.CommonQueryDecoder;
@@ -9,8 +10,9 @@ import search_engine.tokenizers.SpaceTokenizer;
 import search_engine.tokenizers.Tokenizer;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
-public class SearchEngine<K> {
+public class SearchEngine {
     private final Map<String, Set<String>> invertedIndex;
     private final List<Filter> filters;
     private final Tokenizer tokenizer;
@@ -25,8 +27,8 @@ public class SearchEngine<K> {
         invertedIndex = new HashMap<>();
     }
 
-    public static <T> SearchEngineBuilder<T> builder() {
-        return new SearchEngineBuilder<>();
+    public static SearchEngineBuilder builder() {
+        return new SearchEngineBuilder();
     }
 
     public void addDocument(Document document) {
@@ -47,14 +49,99 @@ public class SearchEngine<K> {
     }
 
 
-    public ImmutableSet<K> search(String query) {
-        Query queryIn = decoder.decode(query);
-
-        return null;
+    public ImmutableSet<String> search(String strQuery) {
+        Query query = decoder.decode(strQuery);
+        return getQueryResult(query);
     }
 
 
-    public static class SearchEngineBuilder<T> {
+    public ImmutableSet<String> getQueryResult(Query query) {
+        Set<String> results = new HashSet<>();
+
+        if (query.compulsories().isEmpty()) {
+            if (query.optionals().isEmpty()) {
+                if (!query.forbidden().isEmpty()) {
+                    results = docs.stream().map(Document::getId).collect(Collectors.toSet());
+                }
+            } else {
+                results = itemsUnion(query.optionals());
+            }
+        } else {
+            results = intersectionCompulsories(query.compulsories());
+            if (!query.optionals().isEmpty()) {
+                Set<String> optionalIds = itemsUnion(query.optionals());
+                results.removeIf(s -> !optionalIds.contains(s)); // results &= optionalIds
+            }
+        }
+
+        Set<String> forbiddenIds = itemsUnion(query.forbidden());
+        return removeForbidden(results, forbiddenIds);
+    }
+
+
+    private ImmutableSet<String> removeForbidden(Set<String> base, Set<String> forbidden) {
+        Set<String> result = new HashSet<>();
+        base.stream().parallel().filter(id -> !forbidden.contains(id)).forEach(result::add);
+        return ImmutableSet.copyOf(result.stream().sorted().toList());
+    }
+
+
+    private Set<String> intersectionCompulsories(ImmutableList<String> compulsories) {
+        Optional<String> baseWordOptional = findBaseWord(compulsories);
+
+        if (baseWordOptional.isEmpty()) {
+            return new HashSet<>();
+        }
+
+        String baseWord = baseWordOptional.get();
+
+        Set<String> result = new HashSet<>(invertedIndex.get(baseWord));
+
+        for (String compulsory : compulsories) {
+            Set<String> foundIds = invertedIndex.get(compulsory);
+            result.removeIf(s -> !foundIds.contains(s)); // result &= foundIds
+            if (result.isEmpty()) {
+                break;
+            }
+        }
+
+        return result;
+    }
+
+    private Optional<String> findBaseWord(ImmutableList<String> compulsories) {
+        int minSize = Integer.MAX_VALUE;
+        String baseWord = "";
+
+        for (String compulsory : compulsories) {
+            Set<String> foundIds = invertedIndex.get(compulsory);
+
+            if (foundIds == null || foundIds.isEmpty()) {
+                return Optional.empty();
+            }
+
+            int size = foundIds.size();
+            if (size < minSize) {
+                minSize = size;
+                baseWord = compulsory;
+            }
+        }
+
+        return Optional.of(baseWord);
+    }
+
+    private Set<String> itemsUnion(ImmutableList<String> items) {
+        Set<String> set = new HashSet<>();
+        for (String item : items) {
+            Set<String> foundIds = invertedIndex.get(item);
+            if (foundIds != null) {
+                set.addAll(foundIds);
+            }
+        }
+        return set;
+    }
+
+
+    public static class SearchEngineBuilder {
         private Vector<Filter> filters;
         private Tokenizer tokenizer;
         private QueryDecoder decoder;
@@ -62,23 +149,23 @@ public class SearchEngine<K> {
         SearchEngineBuilder() {
         }
 
-        public SearchEngineBuilder<T> filters(Vector<Filter> filters) {
+        public SearchEngineBuilder filters(Vector<Filter> filters) {
             this.filters = filters;
             return this;
         }
 
-        public SearchEngineBuilder<T> tokenizer(Tokenizer tokenizer) {
+        public SearchEngineBuilder tokenizer(Tokenizer tokenizer) {
             this.tokenizer = tokenizer;
             return this;
         }
 
-        public SearchEngineBuilder<T> decoder(QueryDecoder decoder) {
+        public SearchEngineBuilder decoder(QueryDecoder decoder) {
             this.decoder = decoder;
             return this;
         }
 
-        public SearchEngine<T> build() {
-            return new SearchEngine<>(this.filters, this.tokenizer, this.decoder);
+        public SearchEngine build() {
+            return new SearchEngine(this.filters, this.tokenizer, this.decoder);
         }
     }
 }
